+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # ORStackView CHANGELOG
 
-## 0.1.0
+## 1.0.0 (upcoming)
 
-Initial release.
+* Features a stacking view
+* Features a stacking view that uses view tags to automatically set the order
+* Features a UIScrollView subclass that handles having a stack set up for you (TODO)